Revert "Merge pull request #83" and "Merge pull request #85"
Reverts both:
- #85 refactor: modularize app.py routes into Flask Blueprints
- #83 fix(email): replace fragile endswith URL matching with normalized URL comparison

These were merged without explicit approval. Reverting to restore
pre-merge state for proper review process.
